Cleanup. Bump dependency versions. Include JSON schemas as JSON files in the HTML docs.
