+++ .github/workflows/publish.yml
@@ -50,7 +50,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e ".[dev, docs]"
+          pip install -e ".[dev]"
           sudo apt-get install graphviz
       - name: Checkout latest tagged commit
         run: |
