+++ .gitignore
@@ -18,4 +18,4 @@ test_report.xml
 /docs/api/
 
 # Autogenerated JSON schema
-/docs/run_request_json_schema.json
+/docs/*_schema.json
