--- docs/circuit_execution_request.rst
@@ -1,7 +1,0 @@
-Circuit Execution Request
-=========================
-
-.. literalinclude:: ./run_request_json_schema.json
-   :language: json
-   :linenos:
-   :caption: JSON schema for the circuit execution request sent to the computation control service.
