+++ docs/conf.py
@@ -56,6 +56,9 @@ extensions = [
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
+# Include extra files in the HTML docs.
+html_extra_path = ['run_request_schema.json']
+
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
