+++ docs/json_schemas.rst
@@ -1,0 +1,6 @@
+JSON schemas
+============
+
+The following JSON schemas describe the payload format of the ``iqm-client`` HTTP API.
+
+* `Circuit execution request <run_request_schema.json>`_ sent to the computation control service.
