+++ setup.cfg
@@ -51,10 +51,10 @@ exclude =
 # Add here additional requirements for extra features, to install with:
 # `pip install iqm_client[dev,docs,testing]`
 dev =
-    tox == 3.23.0
+    tox == 3.25.0
 docs =
-    sphinx == 4.2.0
-    sphinx-book-theme == 0.1.5
+    sphinx == 4.5.0
+    sphinx-book-theme == 0.3.2
     gitpython == 3.1.27
 testing =
     gitpython == 3.1.27
@@ -69,5 +69,5 @@ testing =
     jsonschema==4.4.0
     mockito==1.3.0
 cicd =
-    twine >= 3.3.0,<4.0
-    wheel >= 0.36.2,<1.0
+    twine >= 3.3.0, < 4.0
+    wheel >= 0.36.2, < 1.0
