+++ tox.ini
@@ -7,7 +7,7 @@ usedevelop = True
 isolated_build = True
 setenv =
     TOXINIDIR = {toxinidir}
-    VIRTUALENV_PIP=21.3
+    VIRTUALENV_PIP = 22.0.4
 
 [testenv:py{39,310}]
 description =
@@ -30,7 +30,7 @@ setenv =
 extras =
     docs
 commands =
-    python {toxinidir}/docs/generate_run_request_json_schema.py
+    python {toxinidir}/docs/generate_json_schemas.py
     sphinx-build -b {env:BUILD} "{env:DOCSDIR}" "{env:BUILDDIR}/{env:BUILD}" -W
 
 [testenv:build]
